pyproject.toml
[tool.poetry]
name = "betterreads"
version = "0.1.0"
description = ""
authors = ["Ionata <webmaster@ionata.com.au>"]

[tool.poetry.dependencies]
python = "^3.6.1"
celery = { version = "^4.3", extras = ["redis"] }
celery-prometheus-exporter = "^1.7"
django = "^2.2"
django-allauth = "^0.39.1"
django-anymail = { version = "^6.1", extras = ["mailgun"] }
django-axes = "^5.0"
django-celery-beat = "^1.5"
django-celery-results = "^1.1"
django-cors-headers = "^3.1"
django-environ = "^0.4.5"
django-extensions = "^2.2"
django-filter = "^2.2"
django-redis = "~=4.10.0" # does not follow semver
django-rest-auth = "^0.9.5"
django-storages = { version = "^1.7", extras = ["boto3"] }
djangorestframework = "~=3.10.3" # does not follow semver
djangorestframework-filters = ">=1.0.0.dev0"
djangorestframework-jsonapi = "^3.1"
gunicorn = "^19.9"
psycopg2-binary = "^2.8.6"
python-dateutil = "^2.8"
pytz = "*"
Pillow = "^8.1.0"

[tool.poetry.dev-dependencies]
# debugging
django-debug-toolbar = "^2.0"
ipdb = "^0.11.0"
ipython = "^7.2"
ptpython = "^2.0"
# linting
autopep8 = "^1.4"
django-stubs = "^1.5"
mypy = "^0.770"
pycodestyle = "^2.4"
pydocstyle = "^5.0.2"
pylint = "^2.5.2"
pylint-django = "^2.0"
pyls-isort = "^0.1.1"
pyls-mypy = "^0.1.5"
python-language-server = "^0.21.4"
rope = "^0.11.0"
# formatting
black = { version = "^19.10b0", allow-prereleases = true }
isort = { version = "^5.5.4", extras = ["pyproject"] }
pre-commit = "^2.4.0"
# testing
factory_boy = "^2.12.0"
pyhamcrest = "^1.9"
pytest = "^5.3.5"
pytest-django = "^3.8.0"
tblib = "^1.6.0"
# running
werkzeug = "^0.14.1"

[tool.isort]
# this configuration makes isort and black compatible
# https://black.readthedocs.io/en/stable/the_black_code_style.html#how-black-wraps-lines
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
line_length = 88

[tool.pylint.master]
load-plugins = "pylint_django"

[tool.pylint."messages control"]
disable = "bad-continuation,too-many-ancestors,too-few-public-methods"
enable = "useless-suppression"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
